created real screen window, modified Readme to add libraries
