+++ .gitignore
@@ -1,3 +1,5 @@
 .vscode/*
 settings.json
 *.exe
+*.scr
+.test/*
